ch(api_refactor): add LFA api feedbacks
create mochdata
add swagger
separate files in tes

[finishes #168311079]
