+++ .eslintrc.json
@@ -1,5 +1,6 @@
 {
     "env": {
+        "mocha": true,
         "browser": true,
         "es6": true,
         "node": true
