+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 package-lock.json
 .nyc_output
 .env
+UI/image/event_technology.jpg
