+++ server/config/port.js
@@ -1,5 +1,3 @@
 
-// configure our port
-
 const PORT = process.env.PORT || 3000;
 module.exports = PORT;
