+++ server/controllers/mentor_controller.js
@@ -3,7 +3,6 @@ class Mentor_controller
 {
   
 
- //**** mentor accept session start ****/ 
 accept_session=(req,res)=>
 {
 const session_tobe_accepted=db.accept_session(req.params.sessionId);
@@ -14,11 +13,6 @@ else
    return res.status(200).json({status:200,data:{message:"request accepted",user:session_tobe_accepted}})
 }
 }
-//**** ends ****/ 
-
-
-
-//**** mentor reject session start ****/ 
 reject_session=(req,res)=>
 {
 const session_tobe_rejected=db.reject_session(req.params.sessionId);
@@ -29,7 +23,7 @@ else
    return  res.status(200).json({status:200,data:{message:"request rejected",data:session_tobe_rejected}})
 }
 }
-//**** ends ****/ 
+
 
 
 }
