+++ server/index.js
@@ -7,20 +7,23 @@ import mentorroute from './routes/mentor';
 import sessionroute from './routes/sessions';
 import not_found from './routes/default';
 
+const swagger = require('swagger-ui-express');
+const Api_documentation = require('./swagger.json');
+
 
 const app = express();
 
-// routers that will be used
 
 app.use('/api/v1', authroute);
 app.use('/api/v1', adminroute);
 app.use('/api/v1', mentorroute);
 app.use('/api/v1', menteeroute);
 app.use('/api/v1', sessionroute);
+app.use('/documentation', swagger.serve, swagger.setup(Api_documentation));
 
 
 app.use(not_found);
 
-// start server
+
 app.listen(PORT, () => console.log(`server started on PORT ${PORT}`));
 export default app;
