+++ server/models/session_model.js
@@ -3,15 +3,13 @@ const db=user_table.session;
 class Session_modal 
 {
 
-   // select all session request for specific mentor
+  
   
    all_sessions  =(data)=>
    {
    const all_sessions= db.filter(specific=> specific.mentorId==data)
     return all_sessions;
    }
-   
-   // all session made by single user
   
     all_session_request=(data)=>
    {
