+++ server/routes/mentee.js
@@ -6,21 +6,17 @@ import { TOKEN, MENTEE, PASS } from '../middleware/verify';
 const router = Router();
 
 
-// parse application/x-www-form-urlencoded
 router.use(urlencoded({ extended: true }));
 
-// parse application/json
 router.use(json());
 
-// all method that will be accessible by only mentee
 
 router.get('/mentors', TOKEN, MENTEE, mentee.all_mentors);
 router.get('/mentors/:mentorid', TOKEN, MENTEE, mentee.specific_mentor);
 router.post('/sessions', TOKEN, MENTEE, mentee.session_request);
 
-// optional endpoints
 
 router.post('/sessions/:sessionId/review', TOKEN, MENTEE, mentee.review);
 
-// export router
+
 export default router;
